requirements.txt
langchain>=0.1.0
openai==1.12.0
pydantic>=2.5.3
streamlit==1.32.0
python-dotenv>=1.0.1
pandas>=2.2.0
openpyxl==3.1.2
pdfplumber>=0.10.3
python-docx==1.1.0
sqlalchemy==2.0.27
pytest>=8.0.0
langchain-core>=0.1.27
langchain-community==0.0.24
typing-extensions==4.9.0
python-dateutil>=2.8.2
unstructured>=0.6.7
langchain-openai>=0.0.5
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.86.0
google-auth-httplib2>=0.1.0
chromadb>=0.4.22
sentence-transformers>=2.5.1
faiss-cpu>=1.7.4
psycopg2-binary>=2.9.9
supabase>=2.0.0
aiohttp>=3.8.4
asyncio==3.4.3
tiktoken>=0.5.1
pydantic-settings>=2.1.0
PyPDF2>=3.0.0
dateparser 